Update README and refactor autoclicker code
The README file has been updated with more detailed and organized information about the 'RacialClicker' project. Autoclicker logic was refactored using 'SharpHook' library. The aim of this refactor is to improve functionality and readability of the code. Now, any key on the keyboard can serve as a toggle key. Additionally, the autoclicker has a more advanced random delay feature in order to achieve a more realistic clicking rate. Also, the README now includes instructions on running and using the application. Project and solution files have been renamed due to directory restructure.