Reordered introductory content in README.md
The project title "RacialClicker" was moved to the top of README.md. This change was made to give immediate context to the readers about the project. Previously, the title was placed after the project's github badges but it would be more intuitive and user-friendly to have the project title as the first thing the readers see when they open the README.md file.