Update README with detailed installation instructions
The 'Getting Started' section of the README was replaced with a detailed 'Installation Guide'. This was done to better guide users through the process of downloading, installing and running the auto-clicker, specifically addressing certain complexities for MacOS users. Permissions instructions were also added for MacOS users, along with images to aid the process. This is intended to improve users' experience and reduce potential setup issues.